Addressed pr comments
